# docs/contracts/verslomatika-customer-creation.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Private customer business draft creation",
    "version": "0.1.0",
    "description": "Local/test customer draft generation over existing core identity/portfolio. Immutable private revisions; draft_ready is a business proposition and website draft, not public deployment/F1 acceptance. HTML content requires a sandboxed iframe without scripts/same-origin/forms/navigation/network. Request bodies bounded16KiB. Existing v1 auth, intake and operator contracts unchanged."
  },
  "paths": {
    "/customer/v2/creations": {
      "post": {
        "operationId": "startCustomerCreation",
        "security": [
          {
            "CoreSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "202": {
            "description": "Committed safe projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreationViewEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials/session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Local feature disabled or no current ownership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No scoped/public object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Changed idempotency input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Durable action limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Source unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Start"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listCustomerCreations",
        "security": [
          {
            "CoreSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Committed safe projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreationsEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials/session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Local feature disabled or no current ownership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No scoped/public object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Changed idempotency input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Durable action limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Source unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CoreSession": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ]
      },
      "Start": {
        "additionalProperties": false,
        "properties": {
          "portfolio_id": {
            "format": "uuid",
            "title": "Portfolio Id",
            "type": "string"
          },
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "idea": {
            "maxLength": 2500,
            "minLength": 20,
            "title": "Idea",
            "type": "string"
          },
          "canonical_host": {
            "anyOf": [
              {
                "maxLength": 253,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Host"
          },
          "idempotency_key": {
            "format": "uuid",
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "portfolio_id",
          "display_name",
          "idea",
          "canonical_host",
          "idempotency_key"
        ],
        "title": "Start",
        "type": "object"
      },
...
      "Capabilities": {
        "additionalProperties": false,
        "properties": {
          "can_revise": {
            "title": "Can Revise",
            "type": "boolean"
          },
          "can_cancel": {
            "title": "Can Cancel",
            "type": "boolean"
          }
        },
        "required": [
          "can_revise",
          "can_cancel"
        ],
        "title": "Capabilities",
        "type": "object"
      },
      "CreationView": {
        "additionalProperties": false,
        "properties": {
          "creation_id": {
            "format": "uuid",
            "title": "Creation Id",
            "type": "string"
          },
          "portfolio_id": {
            "format": "uuid",
            "title": "Portfolio Id",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "idea": {
            "title": "Idea",
            "type": "string"
          },
          "canonical_host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Host"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "draft_ready",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "stage": {
            "enum": [
              "queued",
              "research",
              "drafting",
              "validation",
              "ready",
              "failed",
              "cancelled"
            ],
            "title": "Stage",
            "type": "string"
          },
          "current_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Revision"
          },
          "active_job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Job Id"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "source_revision": {
            "pattern": "^[a-f0-9]{40}$",
            "title": "Source Revision",
            "type": "string"
          },
          "capabilities": {
            "$ref": "#/components/schemas/Capabilities"
          },
          "latest_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Summary"
          }
        },
        "required": [
          "creation_id",
          "portfolio_id",
          "display_name",
          "idea",
          "canonical_host",
          "status",
          "stage",
          "current_revision",
          "active_job_id",
          "failure_code",
          "created_at",
          "updated_at",
          "source_revision",
          "capabilities",
          "latest_summary"
        ],
        "title": "CreationView",
        "type": "object"
      },
      "Creations": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CreationView"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "Creations",
        "type": "object"
      },
...
      "CreationViewEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "creation.v1"
          },
          "environment": {
            "enum": [
              "local",
              "test"
            ]
          },
          "source_revision": {
            "type": "string"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "data": {
            "$ref": "#/components/schemas/CreationView"
          }
        },
        "required": [
          "contract_version",
          "environment",
          "source_revision",
          "observed_at",
          "request_id",
          "data"
        ]
      },
      "CreationsEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "creation.v1"
          },
          "environment": {
            "enum": [
              "local",
              "test"
            ]
          },
          "source_revision": {
            "type": "string"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "data": {
            "$ref": "#/components/schemas/Creations"
          }
        },
        "required": [
          "contract_version",
          "environment",
          "source_revision",
          "observed_at",
          "request_id",
          "data"
        ]
      },
...
    }
  }
}